+++ src/docker/arcor2_execution_proxy/start.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-
-cd || exit
-./execution_rest_proxy.pex