+++ src/python/arcor2/clients/scene_service.py
@@ -70,7 +70,7 @@ def upsert_collision(model: Models, pose: Pose, mesh_parameters: None | MeshPara
 
     if model.type() == Model3dType.MESH:
 
-        params["meshFileId"] = params.pop("data_id")
+        params["meshFileId"] = params.pop("asset_id")
 
         if mesh_parameters:
             params.update(mesh_parameters.to_dict())
@@ -110,10 +110,10 @@ def delete_all_collisions() -> None:
 
 
 @handle(SceneServiceException, logger, message="Failed to start the scene.")
-def start() -> None:
+def start(timeout: rest.OptTimeout = None) -> None:
     """To be called after all objects are created."""
 
-    rest.call(rest.Method.PUT, f"{URL}/system/start")
+    rest.call(rest.Method.PUT, f"{URL}/system/start", timeout=timeout)
 
 
 @handle(SceneServiceException, logger, message="Failed to stop the scene.")