+++ src/python/arcor2_arserver/robot.py
@@ -4,6 +4,7 @@ from typing import Any
 
 import arcor2.helpers as hlp
 from arcor2.cached import CachedScene
+from arcor2.clients import aio_asset as asset
 from arcor2.data import common
 from arcor2.exceptions import Arcor2Exception
 from arcor2.object_types.abstract import MultiArmRobot, Robot
@@ -11,7 +12,6 @@ from arcor2_arserver import globals as glob
 from arcor2_arserver import logger
 from arcor2_arserver import notifications as notif
 from arcor2_arserver import objects_actions as osa
-from arcor2_arserver.clients import project_service as ps
 from arcor2_arserver.object_types.data import ObjectTypeData
 from arcor2_arserver.object_types.source import function_implemented
 from arcor2_arserver_data import events as sevts
@@ -173,7 +173,7 @@ async def get_robot_meta(obj_type: ObjectTypeData) -> None:
 
     if urdf_name := obj_type.type_def.urdf_package_name:
 
-        if urdf_name not in await ps.files_ids():
+        if not await asset.asset_exists(urdf_name):
             logger.error(f"URDF package {urdf_name} for {obj_type.meta.type} does not exist.")
         else:
             obj_type.robot_meta.urdf_package_filename = urdf_name