+++ src/python/arcor2_arserver/rpc/project.py
@@ -29,7 +29,7 @@ from arcor2_arserver.checks import (
     find_object_action,
 )
 from arcor2_arserver.clients import project_service as storage
-from arcor2_arserver.common import project_names
+from arcor2_arserver.common import project_names, scene_names
 from arcor2_arserver.helpers import (
     ctx_read_lock,
     ctx_write_lock,
@@ -286,7 +286,9 @@ async def add_ap_using_robot_cb(req: srpc.p.AddApUsingRobot.Request, ui: WsClien
         ap = proj.upsert_action_point(common.ActionPoint.uid(), req.args.name, pose.position)
         ori = common.NamedOrientation("default", pose.orientation)
         proj.upsert_orientation(ap.id, ori)
-        joi = common.ProjectRobotJoints("default", req.args.robot_id, joints, True, req.args.arm_id)
+        joi = common.ProjectRobotJoints(
+            "default", req.args.robot_id, joints, True, req.args.arm_id, req.args.end_effector_id
+        )
         proj.upsert_joints(ap.id, joi)
 
         asyncio.ensure_future(notify(ap, ori, joi))
@@ -318,7 +320,9 @@ async def add_action_point_joints_using_robot_cb(
         if req.dry_run:
             return None
 
-        prj = common.ProjectRobotJoints(req.args.name, req.args.robot_id, new_joints, True, req.args.arm_id)
+        prj = common.ProjectRobotJoints(
+            req.args.name, req.args.robot_id, new_joints, True, req.args.arm_id, req.args.end_effector_id
+        )
         proj.upsert_joints(ap.id, prj)
 
         evt = sevts.p.JointsChanged(prj)
@@ -822,6 +826,9 @@ async def new_project_cb(req: srpc.p.NewProject.Request, ui: WsClient) -> None:
 
         unique_name(req.args.name, (await project_names()))
 
+        # TODO workaround for https://gitlab.com/kinalisoft/test-it-off/project/-/issues/16
+        unique_name(req.args.name, (await scene_names()))
+
         if await glob.LOCK.get_write_locks_count() > 1:  # project lock also counts
             raise Arcor2Exception("Project has locked objects")
 
@@ -831,6 +838,11 @@ async def new_project_cb(req: srpc.p.NewProject.Request, ui: WsClient) -> None:
         if glob.LOCK.scene:
             if glob.LOCK.scene.id != req.args.scene_id:
                 raise Arcor2Exception("Another scene is opened.")
+
+            # the scene might not be saved yet - another check
+            # TODO workaround for https://gitlab.com/kinalisoft/test-it-off/project/-/issues/16
+            if req.args.name == glob.LOCK.scene.name:
+                raise Arcor2Exception("A project can't have the same name as the scene.")
 
             if glob.LOCK.scene.has_changes:
                 await save_scene()
@@ -1401,6 +1413,10 @@ async def delete_project_cb(req: srpc.p.DeleteProject.Request, ui: WsClient) -> 
 async def rename_project_cb(req: srpc.p.RenameProject.Request, ui: WsClient) -> None:
 
     unique_name(req.args.new_name, (await project_names()))
+
+    # TODO workaround for https://gitlab.com/kinalisoft/test-it-off/project/-/issues/16
+    unique_name(req.args.new_name, (await scene_names()))
+
     user_name = glob.USERS.user_name(ui)
 
     await ensure_write_locked(req.args.project_id, user_name)