+++ src/python/arcor2_arserver/tests/test_project_const.py
@@ -8,7 +8,7 @@ from arcor2_arserver_data import events, rpc
 from arcor2_arserver_data.client import ARServer, get_id
 
 
-@pytest.mark.additional_deps(["./src.python.arcor2.scripts/upload_builtin_objects.pex"])
+@pytest.mark.additional_deps([["src.python.arcor2.scripts/upload_builtin_objects.pex"]])
 def test_project_const(start_processes: None, ars: ARServer) -> None:
 
     event(ars, events.c.ShowMainScreen)