+++ src/python/arcor2_calibration/tests/test_calibration.py
@@ -6,5 +6,5 @@ from openapi_spec_validator import validate_spec
 
 def test_openapi() -> None:
     validate_spec(
-        yaml.full_load(check_output(["./src.python.arcor2_calibration.scripts/calibration.pex", "--swagger"]))
+        yaml.full_load(check_output(["python", "src.python.arcor2_calibration.scripts/calibration.pex", "--swagger"]))
     )