+++ src/python/arcor2_arserver_data/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
+## [1.0.0] - 2023-02-14
+
+### Changed
+
+- Marked as a stable version.
+
+## [0.22.0] - 2022-12-12
+
+### Changed
+
+- Optional property `end_effector_id` added to `AddActionPointJointsUsingRobot`.
+
 ## [0.21.0] - 2022-10-28
 
 ### Changed