+++ src/python/arcor2_build/CHANGELOG.md
@@ -2,6 +2,20 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
+## [1.0.0] - 2023-02-14
+
+### Changed
+
+- Marked as a stable version.
+- Added dependency on `arcor2_runtime` (so a manually written script can import e.g. `ResourcesException`).
+- Health check end-point changed to `/healthz/ready`.
+
+## [0.25.0] - 2022-12-12
+
+### Changed
+
+- Dependency on `arcor2~=0.26.0`.
+
 ## [0.24.0] - 2022-10-28
 
 ### Changed