+++ src/python/arcor2_runtime/CHANGELOG.md
@@ -2,6 +2,23 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
+## [1.0.0] - 2023-02-14
+
+### Changed
+
+- `ResourcesException` - `__context__` is now set to a first exception occurred during script initialization.
+  - All exceptions that occurred during parallel initialization are still available in the `exceptions` array.
+
+### Added
+
+- `Resources` class has a `scene_start_timeout` argument which is handed over to the Scene client.
+
+## [0.5.0] - 2022-12-12
+
+### Changed
+
+- Dependency on `arcor2~=0.26.0`.
+
 ## [0.4.0] - 2022-10-28
 
 ### Changed